Install WooCommerce as a dev dependency via Composer, and ensure its test framework is available.
Since WooCommerce requires WordPress to be loaded in a certain way, we're better off extending WC_Unit_Test_Case instead of WP_UnitTestCase, since that's the base test case for WooCommerce core.

This commit also includes composer.lock in the repo, as it's designed to be included in version control.