If $wpdb->insert() fails during update_post_meta(), return early from the method.
This will ensure that queries against the $wpdb global will be accurate, and prevent us from attempting to attach post meta when an order row doesn't exist.

Ideally, we'd be returning a WP_Error object (or better yet, throwing an exception), but we're trying to match the method signature of the method being extended; this is really a sign of a larger architectural issue with WooCommerce data stores.