Simplify the installation tests
This commit ensures that the woocommerce_orders table is explicitly dropped at the start of each test (since we're testing that the table can be created). Additionally, some of the methods designed to aid in testing table creation/destruction have been moved from the base TestCase class into this single test case.