Add method_exists check
When setting meta, a method was called using a variable name (name of
meta). However, there may be non-standard meta fields, which cause a
fatal error when an attempt to call a non-existant method is used. To
prevent this there is now a check to ensure the method exists before
it is used.